[team] JSG 마이페이지/관리자 페이지 경로 수정
