+++ teamproject/src/main/java/aihometraining/team/controller/MainController.java
@@ -5,6 +5,7 @@ import javax.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class MainController {
@@ -62,6 +63,36 @@ public class MainController {
 		
 		return "member/wishList";
 		
+	}
+	
+	@GetMapping("/signUpForClass")
+	public String eClassTake(Model model) {
+		
+		
+		
+		model.addAttribute("title", "수강신청");
+		
+		return "eClass/eClassTake";
+	}
+
+	
+	@GetMapping("/payment")
+	public String payment(Model model) {
+		
+		
+		
+		model.addAttribute("title", "결제");
+		
+		return "payment/payment";
+	}
+	
+	@PostMapping("/payment")
+	public String payment(Model model, String a) {
+		
+		
+		model.addAttribute("title", "결제");
+		
+		return "redirect:/mypage/mypaymentList/paymentDetail";
 	}
 
 }
