+++ teamproject/src/main/java/aihometraining/team/controller/RefundController.java
@@ -3,8 +3,10 @@ package aihometraining.team.controller;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
+@RequestMapping("/admin")
 public class RefundController {
 	
 	@GetMapping("/refundList")
