+++ teamproject/src/main/resources/templates/fragments/leftmenu.html
@@ -80,9 +80,9 @@
                     	<span class="fa-angle-right fa right-arrow text-right"></span> 
                     </a>
                       <ul class="nav nav-list tree">
-                        <li><a th:href="@{/paymentList}">#결제 내역(수경)</a></li>
-                        <li><a th:href="@{/refundList}">#환불 내역(수경)</a></li>
-                        <li><a th:href="@{/refundStandard}">#환불 기준 관리(수경)</a></li>
+                        <li><a th:href="@{/admin/paymentList}">#결제 내역(수경)</a></li>
+                        <li><a th:href="@{/admin/refundList}">#환불 내역(수경)</a></li>
+                        <li><a th:href="@{/admin/refundStandard}">#환불 기준 관리(수경)</a></li>
                       </ul>
                     </li>
                     <li class="ripple">
@@ -91,9 +91,9 @@
                     	<span class="fa-angle-right fa right-arrow text-right"></span> 
                     </a>
                       <ul class="nav nav-list tree">
-                        <li><a th:href="@{/givePoint}">#포인트 지급(수경)</a></li>
-                        <li><a th:href="@{/pointList}">#포인트 내역(수경)</a></li>
-                        <li><a th:href="@{/pointStandard}">#포인트 지급 기준(수경)</a></li>
+                        <li><a th:href="@{/admin/givePoint}">#포인트 지급(수경)</a></li>
+                        <li><a th:href="@{/admin/pointList}">#포인트 내역(수경)</a></li>
+                        <li><a th:href="@{/admin/pointStandard}">#포인트 지급 기준(수경)</a></li>
                       </ul>
                     </li>
                     <li class="ripple">
