+++ teamproject/src/main/resources/templates/payment/paymentList.html
@@ -186,7 +186,7 @@
 							<tr>
 								<td>2022-03-23</td>
 								<!-- (paymentCode=${l.paymentCode}) -->
-								<td><a th:href="@{/paymentList/paymentDetail(paymentCode=)}">e20220323_00001</a></td>
+								<td><a th:href="@{/admin/paymentList/paymentDetail(paymentCode=)}">e20220323_00001</a></td>
 								<td>운동클래스</td>
 								<td>id005@email.com</td>
 								<td>카드</td>
@@ -197,7 +197,7 @@
 							</tr>
 							<tr>
 								<td>2022-02-23</td>
-								<td><a th:href="@{/paymentList/paymentDetail(paymentCode=)}">c20220223_00001</a></td>
+								<td><a th:href="@{/admin/paymentList/paymentDetail(paymentCode=)}">c20220223_00001</a></td>
 								<td>챌린지</td>
 								<td>id005@email.com</td>
 								<td>현금</td>
@@ -208,7 +208,7 @@
 							</tr>
 							<tr>
 								<td>2022-02-23</td>
-								<td><a th:href="@{/paymentList/paymentDetail(paymentCode=)}">c20220223_00002</a></td>
+								<td><a th:href="@{/admin/paymentList/paymentDetail(paymentCode=)}">c20220223_00002</a></td>
 								<td>챌린지</td>
 								<td>id006@email.com</td>
 								<td>현금</td>
@@ -219,7 +219,7 @@
 							</tr>
 							<tr>
 								<td>2022-02-24</td>
-								<td><a th:href="@{/paymentList/paymentDetail(paymentCode=)}">e20220224_00001</a></td>
+								<td><a th:href="@{/admin/paymentList/paymentDetail(paymentCode=)}">e20220224_00001</a></td>
 								<td>운동클래스</td>
 								<td>id006@email.com</td>
 								<td>카드</td>
@@ -230,7 +230,7 @@
 							</tr>
 							<tr>
 								<td>2022-02-25</td>
-								<td><a th:href="@{/paymentList/paymentDetail(paymentCode=)}">c20220225_00001</a></td>
+								<td><a th:href="@{/admin/paymentList/paymentDetail(paymentCode=)}">c20220225_00001</a></td>
 								<td>챌린지</td>
 								<td>id007@email.com</td>
 								<td>현금</td>
@@ -241,7 +241,7 @@
 							</tr>
 							<tr>
 								<td>2022-02-22</td>
-								<td><a th:href="@{/paymentList/paymentDetail(paymentCode=)}">e20220222_00001</a></td>
+								<td><a th:href="@{/admin/paymentList/paymentDetail(paymentCode=)}">e20220222_00001</a></td>
 								<td>운동클래스</td>
 								<td>id007@email.com</td>
 								<td>현금</td>
