+++ teamproject/src/main/resources/templates/refund/refundList.html
@@ -266,7 +266,7 @@ input[name="searchValue"] {
 									<tr>
 										<td>2022-03-23</td>
 										<td>2022-03-23</td>
-										<td><a th:href="@{/refundList/refundDetail(refundCode=)}">e20220224_00001</a></td>
+										<td><a th:href="@{/admin/refundList/refundDetail(refundCode=)}">e20220224_00001</a></td>
 										<td>운동클래스</td>
 										<td>홍길동</td>
 										<td>
@@ -276,7 +276,7 @@ input[name="searchValue"] {
 									<tr>
 										<td>2022-02-23</td>
 										<td>2022-02-23</td>
-										<td><a th:href="@{/refundList/refundDetail(refundCode=)}">c20220223_00002</a></td>
+										<td><a th:href="@{/admin/refundList/refundDetail(refundCode=)}">c20220223_00002</a></td>
 										<td>챌린지</td>
 										<td>홍길동</td>
 										<td>
@@ -286,7 +286,7 @@ input[name="searchValue"] {
 									<tr>
 										<td>2022-02-23</td>
 										<td>2022-02-23</td>
-										<td><a th:href="@{/refundList/refundDetail(refundCode=)}">c20220223_00002</a></td>
+										<td><a th:href="@{/admin/refundList/refundDetail(refundCode=)}">c20220223_00002</a></td>
 										<td>챌린지</td>
 										<td>이순신</td>
 										<td>
@@ -296,7 +296,7 @@ input[name="searchValue"] {
 									<tr>
 										<td>2022-02-24</td>
 										<td>2022-02-24</td>
-										<td><a th:href="@{/refundList/refundDetail(refundCode=)}">e20220224_00001</a></td>
+										<td><a th:href="@{/admin/refundList/refundDetail(refundCode=)}">e20220224_00001</a></td>
 										<td>운동클래스</td>
 										<td>이순신</td>
 										<td>
